updated template format
